requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-google-maps==0.13.0
django-rename-app==0.1.7
djangorestframework==3.14.0
gurobipy==11.0.2
idna==3.6
numpy==2.0.0
pandas==2.2.2
psycopg==3.1.18
psycopg-binary==3.1.18
python-dateutil==2.9.0.post0
pytz==2024.1
razorpay==1.4.1
redis==5.0.3
requests==2.31.0
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
